Split more oversized test fixtures
