+++ scripts/file-size-hygiene-baseline.txt
@@ -12,10 +12,10 @@
 950083 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionViewModel.kt
 278779 app/src/test/java/com/pocketshell/app/composer/PromptComposerViewModelTest.kt
 138541 app/src/test/java/com/pocketshell/app/session/AgentConversationRepositoryTest.kt
-738646 app/src/test/java/com/pocketshell/app/tmux/TmuxSessionViewModelTest.kt
+728066 app/src/test/java/com/pocketshell/app/tmux/TmuxSessionViewModelTest.kt
 1146350 docs/mockups/feedback/folder-tree-target-20260604.png
 182790 docs/screenshots/readme-settings.png
 194363 scripts/ci-journey-suite.sh
 160115 shared/core-tmux/src/main/java/com/pocketshell/core/tmux/TmuxClient.kt
-171240 shared/core-tmux/src/test/java/com/pocketshell/core/tmux/TmuxClientTest.kt
-153076 shared/ui-kit/src/test/java/com/pocketshell/uikit/render/DesignRenders.kt
+152167 shared/core-tmux/src/test/java/com/pocketshell/core/tmux/TmuxClientTest.kt
+144031 shared/ui-kit/src/test/java/com/pocketshell/uikit/render/DesignRenders.kt
